Edge to edge: handle SleepActivity
See commit 4550d5cbfb13e14d82e75e74d7cd4b9e5be7c51f (Enable edge-to-edge
mode and fix MainActivity, 2025-10-31) for more details, this does the
same improvements which were done for MainActivity already.
